+++ components/search/src/configuration_types.rs
@@ -52,6 +52,10 @@ pub(crate) struct JSONEngineUrl {
     /// appended to the end of the query. This may be skipped if `{searchTerms}`
     /// is included in the base.
     pub search_term_param_name: Option<String>,
+
+    /// The display name of the URL, if any. This is useful if the URL
+    /// corresponds to a brand name distinct from the engine's brand name.
+    pub display_name: Option<String>,
 }
 
 /// Reflects `types::SearchEngineUrls`, but using `EngineUrl`.
@@ -69,6 +73,9 @@ pub(crate) struct JSONEngineUrls {
 
     /// The URL of the search engine homepage.
     pub search_form: Option<JSONEngineUrl>,
+
+    /// The URL to use for visual searches.
+    pub visual_search: Option<JSONEngineUrl>,
 }
 
 /// Represents the engine base section of the configuration.