+++ components/search/src/filter.rs
@@ -22,6 +22,7 @@ impl From<JSONEngineUrl> for SearchEngineUrl {
             method: url.method.unwrap_or_default().as_str().to_string(),
             params: url.params.unwrap_or_default(),
             search_term_param_name: url.search_term_param_name,
+            display_name: url.display_name,
         }
     }
 }
@@ -40,6 +41,9 @@ impl SearchEngineUrl {
         if let Some(search_term_param_name) = &preferred.search_term_param_name {
             self.search_term_param_name = Some(search_term_param_name.clone());
         }
+        if let Some(display_name) = &preferred.display_name {
+            self.display_name = Some(display_name.clone());
+        }
     }
 }
 
@@ -50,6 +54,7 @@ impl From<JSONEngineUrls> for SearchEngineUrls {
             suggestions: urls.suggestions.map(|suggestions| suggestions.into()),
             trending: urls.trending.map(|trending| trending.into()),
             search_form: urls.search_form.map(|search_form| search_form.into()),
+            visual_search: urls.visual_search.map(|visual_search| visual_search.into()),
         }
     }
 }
@@ -75,6 +80,12 @@ impl SearchEngineUrls {
             match &mut self.search_form {
                 Some(search_form) => search_form.merge(search_form_url),
                 None => self.search_form = Some(search_form_url.clone().into()),
+            };
+        }
+        if let Some(visual_search_url) = &preferred.visual_search {
+            match &mut self.visual_search {
+                Some(visual_search) => visual_search.merge(visual_search_url),
+                None => self.visual_search = Some(visual_search_url.clone().into()),
             };
         }
     }
@@ -490,6 +501,7 @@ mod tests {
                     method: None,
                     params: None,
                     search_term_param_name: None,
+                    display_name: None,
                 }),
                 ..Default::default()
             },
@@ -535,10 +547,12 @@ mod tests {
                         method: None,
                         params: None,
                         search_term_param_name: None,
+                        display_name: None,
                     }),
                     suggestions: None,
                     trending: None,
                     search_form: None,
+                    visual_search: None,
                 },
             },
             &JSONEngineVariant {
@@ -575,10 +589,12 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
+                        display_name: None,
                     },
                     suggestions: None,
                     trending: None,
-                    search_form: None
+                    search_form: None,
+                    visual_search: None,
                 },
                 click_url: None
             }
@@ -610,6 +626,7 @@ mod tests {
                     },
                 ]),
                 search_term_param_name: Some("baz".to_string()),
+                display_name: None,
             }),
             suggestions: Some(JSONEngineUrl {
                 base: Some("https://example.com/suggestions".to_string()),
@@ -621,6 +638,7 @@ mod tests {
                     experiment_config: Some("suggest-experiment-value".to_string()),
                 }]),
                 search_term_param_name: Some("suggest".to_string()),
+                display_name: None,
             }),
             trending: Some(JSONEngineUrl {
                 base: Some("https://example.com/trending".to_string()),
@@ -632,6 +650,7 @@ mod tests {
                     experiment_config: None,
                 }]),
                 search_term_param_name: None,
+                display_name: None,
             }),
             search_form: Some(JSONEngineUrl {
                 base: Some("https://example.com/search_form".to_string()),
@@ -643,6 +662,19 @@ mod tests {
                     experiment_config: None,
                 }]),
                 search_term_param_name: None,
+                display_name: None,
+            }),
+            visual_search: Some(JSONEngineUrl {
+                base: Some("https://example.com/visual_search".to_string()),
+                method: Some(crate::JSONEngineMethod::Get),
+                params: Some(vec![SearchUrlParam {
+                    name: "visual-search-name".to_string(),
+                    value: Some("visual-search-value".to_string()),
+                    enterprise_value: None,
+                    experiment_config: None,
+                }]),
+                search_term_param_name: Some("url".to_string()),
+                display_name: Some("Visual Search".to_string()),
             }),
         },
     });
@@ -699,6 +731,7 @@ mod tests {
                             },
                         ],
                         search_term_param_name: Some("baz".to_string()),
+                        display_name: None,
                     },
                     suggestions: Some(SearchEngineUrl {
                         base: "https://example.com/suggestions".to_string(),
@@ -710,6 +743,7 @@ mod tests {
                             experiment_config: Some("suggest-experiment-value".to_string()),
                         }],
                         search_term_param_name: Some("suggest".to_string()),
+                        display_name: None,
                     }),
                     trending: Some(SearchEngineUrl {
                         base: "https://example.com/trending".to_string(),
@@ -721,6 +755,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: None,
+                        display_name: None,
                     }),
                     search_form: Some(SearchEngineUrl {
                         base: "https://example.com/search_form".to_string(),
@@ -732,6 +767,19 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: None,
+                        display_name: None,
+                    }),
+                    visual_search: Some(SearchEngineUrl {
+                        base: "https://example.com/visual_search".to_string(),
+                        method: "GET".to_string(),
+                        params: vec![SearchUrlParam {
+                            name: "visual-search-name".to_string(),
+                            value: Some("visual-search-value".to_string()),
+                            enterprise_value: None,
+                            experiment_config: None,
+                        }],
+                        search_term_param_name: Some("url".to_string()),
+                        display_name: Some("Visual Search".to_string()),
                     }),
                 },
                 click_url: None
@@ -764,6 +812,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("ship".to_string()),
+                        display_name: None,
                     }),
                     suggestions: Some(JSONEngineUrl {
                         base: Some("https://example.com/suggestions-variant".to_string()),
@@ -775,6 +824,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("variant".to_string()),
+                        display_name: None,
                     }),
                     trending: Some(JSONEngineUrl {
                         base: Some("https://example.com/trending-variant".to_string()),
@@ -786,6 +836,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("trend".to_string()),
+                        display_name: None,
                     }),
                     search_form: Some(JSONEngineUrl {
                         base: Some("https://example.com/search_form".to_string()),
@@ -797,6 +848,19 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: None,
+                        display_name: None,
+                    }),
+                    visual_search: Some(JSONEngineUrl {
+                        base: Some("https://example.com/visual-search-variant".to_string()),
+                        method: Some(JSONEngineMethod::Get),
+                        params: Some(vec![SearchUrlParam {
+                            name: "visual-search-variant-name".to_string(),
+                            value: Some("visual-search-variant-value".to_string()),
+                            enterprise_value: None,
+                            experiment_config: None,
+                        }]),
+                        search_term_param_name: Some("url_variant".to_string()),
+                        display_name: Some("Visual Search Variant".to_string()),
                     }),
                 }),
                 sub_variants: vec![],
@@ -828,6 +892,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("ship".to_string()),
+                        display_name: None,
                     },
                     suggestions: Some(SearchEngineUrl {
                         base: "https://example.com/suggestions-variant".to_string(),
@@ -839,6 +904,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("variant".to_string()),
+                        display_name: None,
                     }),
                     trending: Some(SearchEngineUrl {
                         base: "https://example.com/trending-variant".to_string(),
@@ -850,6 +916,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("trend".to_string()),
+                        display_name: None,
                     }),
                     search_form: Some(SearchEngineUrl {
                         base: "https://example.com/search_form".to_string(),
@@ -861,6 +928,19 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: None,
+                        display_name: None,
+                    }),
+                    visual_search: Some(SearchEngineUrl {
+                        base: "https://example.com/visual-search-variant".to_string(),
+                        method: "GET".to_string(),
+                        params: vec![SearchUrlParam {
+                            name: "visual-search-variant-name".to_string(),
+                            value: Some("visual-search-variant-value".to_string()),
+                            enterprise_value: None,
+                            experiment_config: None,
+                        }],
+                        search_term_param_name: Some("url_variant".to_string()),
+                        display_name: Some("Visual Search Variant".to_string()),
                     }),
                 },
                 click_url: None
@@ -893,6 +973,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("ship".to_string()),
+                        display_name: None,
                     }),
                     suggestions: Some(JSONEngineUrl {
                         base: Some("https://example.com/suggestions-variant".to_string()),
@@ -904,6 +985,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("variant".to_string()),
+                        display_name: None,
                     }),
                     trending: Some(JSONEngineUrl {
                         base: Some("https://example.com/trending-variant".to_string()),
@@ -915,6 +997,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("trend".to_string()),
+                        display_name: None,
                     }),
                     search_form: Some(JSONEngineUrl {
                         base: Some("https://example.com/search-form-variant".to_string()),
@@ -926,6 +1009,19 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: None,
+                        display_name: None,
+                    }),
+                    visual_search: Some(JSONEngineUrl {
+                        base: Some("https://example.com/visual-search-variant".to_string()),
+                        method: Some(JSONEngineMethod::Get),
+                        params: Some(vec![SearchUrlParam {
+                            name: "visual-search-variant-name".to_string(),
+                            value: Some("visual-search-variant-value".to_string()),
+                            enterprise_value: None,
+                            experiment_config: None,
+                        }]),
+                        search_term_param_name: Some("url_variant".to_string()),
+                        display_name: Some("Visual Search Variant".to_string()),
                     }),
                 }),
                 // This would be the list of sub-variants for this part of the
@@ -954,6 +1050,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("shuttle".to_string()),
+                        display_name: None,
                     }),
                     suggestions: Some(JSONEngineUrl {
                         base: Some("https://example.com/suggestions-subvariant".to_string()),
@@ -965,6 +1062,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("subvariant".to_string()),
+                        display_name: None,
                     }),
                     trending: Some(JSONEngineUrl {
                         base: Some("https://example.com/trending-subvariant".to_string()),
@@ -976,6 +1074,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("subtrend".to_string()),
+                        display_name: None,
                     }),
                     search_form: Some(JSONEngineUrl {
                         base: Some("https://example.com/search-form-subvariant".to_string()),
@@ -987,6 +1086,19 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: None,
+                        display_name: None,
+                    }),
+                    visual_search: Some(JSONEngineUrl {
+                        base: Some("https://example.com/visual-search-subvariant".to_string()),
+                        method: Some(JSONEngineMethod::Get),
+                        params: Some(vec![SearchUrlParam {
+                            name: "visual-search-subvariant-name".to_string(),
+                            value: Some("visual-search-subvariant-value".to_string()),
+                            enterprise_value: None,
+                            experiment_config: None,
+                        }]),
+                        search_term_param_name: Some("url_subvariant".to_string()),
+                        display_name: Some("Visual Search Subvariant".to_string()),
                     }),
                 }),
                 sub_variants: vec![],
@@ -1017,6 +1129,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("shuttle".to_string()),
+                        display_name: None,
                     },
                     suggestions: Some(SearchEngineUrl {
                         base: "https://example.com/suggestions-subvariant".to_string(),
@@ -1028,6 +1141,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("subvariant".to_string()),
+                        display_name: None,
                     }),
                     trending: Some(SearchEngineUrl {
                         base: "https://example.com/trending-subvariant".to_string(),
@@ -1039,6 +1153,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("subtrend".to_string()),
+                        display_name: None,
                     }),
                     search_form: Some(SearchEngineUrl {
                         base: "https://example.com/search-form-subvariant".to_string(),
@@ -1050,6 +1165,19 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: None,
+                        display_name: None,
+                    }),
+                    visual_search: Some(SearchEngineUrl {
+                        base: "https://example.com/visual-search-subvariant".to_string(),
+                        method: "GET".to_string(),
+                        params: vec![SearchUrlParam {
+                            name: "visual-search-subvariant-name".to_string(),
+                            value: Some("visual-search-subvariant-value".to_string()),
+                            enterprise_value: None,
+                            experiment_config: None,
+                        }],
+                        search_term_param_name: Some("url_subvariant".to_string()),
+                        display_name: Some("Visual Search Subvariant".to_string()),
                     }),
                 },
                 click_url: None
@@ -1068,6 +1196,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
+                        display_name: None,
                     },
                     ..Default::default()
                 },
@@ -1082,6 +1211,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
+                        display_name: None,
                     },
                     ..Default::default()
                 },
@@ -1096,6 +1226,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
+                        display_name: None,
                     },
                     ..Default::default()
                 },
@@ -1110,6 +1241,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
+                        display_name: None,
                     },
                     ..Default::default()
                 },